+++ src/com/adgoji/mollie/client.clj
@@ -83,13 +83,13 @@
        (cond-> body
          :always         (response-transformer)
          check-response? (spec/check spec))
-       (let [anomaly (case status
-                       (400 401 422) ::anomalies/incorrect
-                       403           ::anomalies/forbidden
-                       (404 410)     ::anomalies/not-found
-                       405           ::anomalies/unsupported
-                       500           ::anomalies/fault
-                       503           ::anomalies/busy)
+       (let [anomaly (case (long status)
+                       (400 401 409 422) ::anomalies/incorrect
+                       403               ::anomalies/forbidden
+                       (404 410)         ::anomalies/not-found
+                       405               ::anomalies/unsupported
+                       500               ::anomalies/fault
+                       503               ::anomalies/busy)
              error   {::anomalies/category anomaly
                       :error               body}]
          (if throw-exceptions?